sa, sb deux elems

+++ includes/push_swap.h
@@ -20,12 +20,12 @@ typedef struct	s_stacks
 	t_list		*sa;
 	t_list		*bl_sa;		//before last
 	t_list		*l_sa;		//last
-	size_t		nb_sa;		//nb d'elements
+	int			nb_sa;		//nb d'elements
 
 	t_list		*sb;
 	t_list		*bl_sb;		//before last
 	t_list		*l_sb;		//last
-	size_t		nb_sb;		//nb d'elements
+	int			nb_sb;		//nb d'elements
 
 }				t_stacks;
